Fix indicator can't disappear in some cases
When an app A starts recording and then passed to other process itself. A became the accessor and receiver one after another, and the indicator appeared because of accessor. but when the recording ended, the accessor was removed first and the callback only receive the notify of receiver. So the indicator can't disappear in addition to reboot.

Change-Id: I1fe5eb2dcb216ab18907ecbcb3a58edf27b3c21f
Signed-off-by: Dmitrii <bankersenator@gmail.com>